requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
anthropic>=0.39.0
pandas>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.9.0
